+++ http_testing/cookie.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 from typing import Optional, Union
 
 from attrs import define
@@ -12,4 +13,4 @@ class Cookie:
     domain: Optional[str] = None
     path: Optional[str] = None
     secure: Optional[bool] = None
-    expires: Optional[int] = None
+    expires: Optional[datetime] = None